requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
schedule==1.1.0
pyyaml==6.0
python-dotenv==0.21.0 